Add the ability to launch with Extras
Added ParseTypes class to convert basic JSON types to Java types that
can go in a Bundle. Plugin looks for extras object being passed in.
Added ability to get data back from launched application, if available.
Changed the Javascript call to allow success and error callbacks to be
passed as part of the options object, if desired.